requirements.txt
flask
openai
pymongo
python-dotenv
docx2txt
pdfminer.six
werkzeug